fix(landing): restrict mobile nav header to landing page (#2685)
* kni landing page first round of development

* external links updated from copy doc

* Landing page copy doc and PR comment updates

* removed unused useState functions from dummy data

* updated local hrefs with routes, removed debug code

* landing page QA updates

* scroll into view functionality added to content nav links

* removing redundant header from landing page template. header remaining in layout template.

* updating mobile hero image positioning

* fixing typescript error on useRef

* mobile header styles and mobile nav

* styling updates. data-test-id added to homepage logo

* adding interfaces for data-test-id props on Homepage Link and Logo components

* replace mobile HomepageLink component with Link and Image

* react-tweet-embed package added. replacing twitter screenshots with embedded links. updating content security policy to allow twitter.

* fixing box shadows and overflow on content images

* updating and improving mobile styles

* updating site so that mobile header is only active on landing page.
