+++ frontend/src/components/Header/style.ts
@@ -11,13 +11,7 @@ export const Wrapper = styled.div`
   position: fixed;
   top: 0;
   width: 100%;
-  z-index: 2;
-
-  @media (max-width: 768px) {
-    height: 100%;
-    position: relative;
-    display: flex;
-  }
+  z-index: 1;
 `;
 
 export const MainWrapper = styled.div`
@@ -26,98 +20,6 @@ export const MainWrapper = styled.div`
   height: inherit; /* Take up full height of parent. */
   justify-content: space-between;
   padding: 0 16px;
-
-  @media (max-width: 768px) {
-    flex-direction: column;
-    justify-content: flex-start;
-    padding-top: ${HEADER_HEIGHT_PX}px;
-  }
-`;
-
-export const MobileHomeLink = styled.span`
-  display: none;
-  @media (max-width: 768px) {
-    display: block;
-  }
-`;
-
-export const DesktopHomeLink = styled.span`
-  @media (max-width: 768px) {
-    display: none;
-  }
-`;
-
-export const MobileMenuButton = styled.div`
-  display: none;
-
-  @media (max-width: 768px) {
-    display: block;
-    color: white;
-    cursor: pointer;
-    position: relative;
-    z-index: 3;
-    overflow: hidden;
-  }
-`;
-
-export const MobileMenuButtonBar = styled.div`
-  display: none;
-
-  @media (max-width: 768px) {
-    display: block;
-    width: 30px;
-    height: 2px;
-    background-color: #fff;
-    margin: 6px 0;
-    transition: 0.4s;
-
-    &.open {
-      &:first-of-type {
-        transform: translateY(8px) rotate(45deg);
-      }
-
-      &:nth-of-type(2) {
-        transform: translateX(60px);
-      }
-
-      &:nth-of-type(3) {
-        transform: translateY(-8px) rotate(-45deg);
-      }
-    }
-  }
-`;
-
-export const MobileNavTray = styled.div`
-  @media (max-width: 768px) {
-    position: absolute;
-    transform: translateX(100vw);
-    right: 0;
-    top: 0;
-    width: 80%;
-    height: 100vh;
-    transition: transform 0.35s ease-out;
-
-    &.active {
-      transform: translateX(0);
-    }
-  }
-`;
-
-export const MobileNavWrapper = styled.div`
-  position: relative;
-  background-color: ${PT_TEXT_COLOR};
-  width: 100%;
-  height: ${HEADER_HEIGHT_PX}px;
-
-  @media (max-width: 768px) {
-    position: fixed;
-    z-index: 3;
-    top: 0;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 16px;
-  }
 `;
 
 export const Left = styled.span`
@@ -128,36 +30,17 @@ export const Left = styled.span`
   a {
     display: flex; /* Ensures the anchor wrapping the logo has correct line height. */
   }
-
-  @media (max-width: 768px) {
-    flex-direction: column;
-    align-items: flex-start;
-  }
 `;
 
 export const Right = styled.span`
   align-items: center;
   display: flex;
   gap: 24px;
-
-  @media (max-width: 768px) {
-    flex-direction: column;
-    align-items: flex-start;
-  }
 `;
 
 export const Nav = styled.span`
   display: flex;
   gap: 16px;
-
-  @media (max-width: 768px) {
-    flex-direction: column;
-    align-items: flex-start;
-    padding-top: 16px;
-    position: relative;
-    left: -14px;
-    margin-bottom: 26px;
-  }
 `;
 
 const button = css`
@@ -222,16 +105,4 @@ export const BetaChip = styled(Chip)`
   color: white;
   margin-left: ${PT_GRID_SIZE_PX / 2}px;
   height: ${PT_GRID_SIZE_PX * 2}px !important;
-`;
-
-export const HiringLink = styled.a`
-  background-color: #3867fa;
-  padding: 7px 14px;
-  border-radius: 4px;
-  color: #fff;
-  font-weight: 600;
-  &:hover {
-    color: #fff;
-    text-decoration: none;
-  }
 `;
