--- frontend/src/components/LandingHeader/index.module.scss
@@ -1,55 +1,0 @@
-// HEADER + NAV
-.header {
-  background-color: #000;
-  border-bottom: 1px solid #444444;
-  height: 50px;
-  display: flex;
-
-  &LogoContainer {
-    border-right: 1px solid #444444;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-    max-width: 150px;
-    padding: 10px 26px;
-  }
-
-  &NavContainer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    padding: 0 40px;
-    z-index: 1;
-
-    a {
-      color: #fff;
-      font-size: 13px;
-      line-height: 20px;
-      letter-spacing: 0.3px;
-      font-weight: 600;
-      transition: all 0.35s;
-
-      &:hover {
-        text-decoration: none;
-        color: #3867fa;
-      }
-
-      & + a {
-        margin-left: 30px;
-      }
-
-      &.btnLink {
-        background-color: #3867fa;
-        padding: 7px 14px;
-        border-radius: 4px;
-
-        &:hover {
-          color: #fff;
-        }
-      }
-    }
-  }
-}
-// END HEADER + NAV
